feat DOUBLE保留小数点后两位

+++ src/main/java/fund/handler/CalDataHandler.java
@@ -47,7 +47,7 @@ public class CalDataHandler extends AbstractFundBeanHandler {
 
         // 基金存续时间
         long totalDay = TimeUtil.calYearBetween(startDay.getDate(), endDay.getDate()) + 1;
-        bean.setDurationDay((int) totalDay);
+        bean.setDurationDay(totalDay);
         int tradingDay = dayList.size();
 
         // 最新一日相关信息
